refactor: move Months component to Months.tsx file
