+++ src/components/Certificates.tsx
@@ -3,43 +3,7 @@ import { TextField } from "./TextField";
 import { TextArea } from "./TextArea";
 import { CircleButton } from "./Buttons";
 import { useOpenStatus } from "../main";
-
-const Months = ({ isDisabled }: { isDisabled: boolean }) => {
-  const months = [
-    "Month",
-    "January",
-    "February",
-    "March",
-    "April",
-    "May",
-    "June",
-    "July",
-    "August",
-    "September",
-    "October",
-    "November",
-    "December",
-  ];
-
-  return (
-    <>
-      <label htmlFor="months" className="block w-full">
-        <select
-          name="months"
-          id="months"
-          disabled={isDisabled}
-          className="rounded-lg bg-gray-100 border-none min-w-full"
-        >
-          {months.map((val, idx) => (
-            <option value={val} key={idx}>
-              {val}
-            </option>
-          ))}
-        </select>
-      </label>
-    </>
-  );
-};
+import { Months } from "./Months";
 
 const Years = ({ isDisabled }: { isDisabled: boolean }) => {
   const years = [];
